+++ osrframework/wrappers/badoo.py
@@ -97,7 +97,7 @@ class Badoo(Platform):
         # Definition of regular expressions to be searched in usufy mode
         self.fieldsRegExp["usufy"] = {}
         # Example of fields:
-        #self.fieldsRegExp["usufy"]["i3visio.location"] = ""
+        self.fieldsRegExp["usufy"]["i3visio.location"] = {"start": '<span class="js-location-label">', "end": '</span>'}
 
         # Definition of regular expressions to be searched in searchfy mode
         #self.fieldsRegExp["searchfy"] = {}